+++ spec/lib/session_encryptor_spec.rb
@@ -47,27 +47,6 @@ RSpec.describe SessionEncryptor do
       expect(result.fetch('warden.user.user.session')['encrypted_pii']).to_not eq nil
     end
 
-    it 'can decrypt PII bundle with Pii::Cacher' do
-      session = { 'warden.user.user.session' => {
-        'decrypted_pii' => { 'ssn' => '666-66-6666' }.to_json,
-      } }
-
-      ciphertext = subject.dump(session)
-
-      result = subject.load(ciphertext)
-      pii_cacher = Pii::Cacher.new(nil, result.fetch('warden.user.user.session'))
-
-      expect(result.fetch('warden.user.user.session')['decrypted_pii']).to eq nil
-      expect(result.fetch('warden.user.user.session')['encrypted_pii']).to_not eq nil
-
-      pii_cacher.fetch
-      expect(JSON.parse(result.fetch('warden.user.user.session')['decrypted_pii'])).to eq(
-        {
-          'ssn' => '666-66-6666',
-        },
-      )
-    end
-
     it 'KMS encrypts/decrypts doc auth elements of the session' do
       session = { 'warden.user.user.session' => {
         'idv' => { 'ssn' => '666-66-6666' },
